+++ src/main/java/com/rymcu/forest/mapper/RoleMapper.java
@@ -12,7 +12,7 @@ public interface RoleMapper extends Mapper<Role> {
 
     Role selectRoleByInputCode(@Param("inputCode") String inputCode);
 
-    Integer updateStatus(@Param("idRole") Integer idRole, @Param("status") String status);
+    Integer updateStatus(@Param("idRole") Long idRole, @Param("status") String status);
 
-    Integer update(@Param("idRole") Integer idRole, @Param("name") String name, @Param("inputCode") String inputCode, @Param("weights") Integer weights);
+    Integer update(@Param("idRole") Long idRole, @Param("name") String name, @Param("inputCode") String inputCode, @Param("weights") Integer weights);
 }
